api/model/requirements.txt
bertopic
fastapi
uvicorn
gunicorn
httpx
numpy
pandas
python-dotenv
nltk
SQLAlchemy
pymongo
passlib
beautifulsoup4
python-jose
python-multipart
openai
qdrant_client
mlflow
langchain
psutil